Sprint 3 Complete

+++ FYP/AnimationManager.cpp
@@ -1,9 +1,28 @@
 #include "AnimationManager.h"
 
+static bool instanceFlag = false;
+static AnimationManager* instance = NULL;
+
 AnimationManager::AnimationManager()
 {
 
+	checkDestroyed = false;
+
 	srand(time(NULL));
+
+	if (!smokeEffectTexture.loadFromFile("smokeTrail9.png"))
+	{
+
+	}
+
+	smokeEffectTexture.setSmooth(true);
+
+	for (int i = 0; i < 20; i++)
+	{
+		smokeEffect.addFrame(sf::IntRect(0, 12 * i, 12, 12));
+	}
+
+	smokeEffect.setSpriteSheet(smokeEffectTexture);
 
 	if (!largeExplosionTexture.loadFromFile("explosionSheet3.png"))
 	{
@@ -26,31 +45,6 @@ AnimationManager::AnimationManager()
 	}
 
 	largeExplosion.setSpriteSheet(largeExplosionTexture);
-	//largeExplosion.addFrame(sf::IntRect(0, 0, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(64, 0, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(128, 0, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(192, 0, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(256, 0, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(0, 64, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(64, 64, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(128, 64, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(192, 64, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(256, 64, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(0, 128, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(64, 128, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(128, 128, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(192, 128, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(256, 128, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(0, 192, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(64, 192, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(128, 192, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(192, 192, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(256, 192, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(0, 256, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(64, 256, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(128, 256, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(192, 256, 64, 64));
-	//largeExplosion.addFrame(sf::IntRect(256, 256, 64, 64));
 
 	if (!smallExplosionTexture.loadFromFile("explosionSheet3.png"))
 	{
@@ -72,33 +66,6 @@ AnimationManager::AnimationManager()
 		smallExplosion.addFrame(sf::IntRect(600, 100 * i, 100, 100));
 		smallExplosion.addFrame(sf::IntRect(700, 100 * i, 100, 100));
 		smallExplosion.addFrame(sf::IntRect(800, 100 * i, 100, 100));
-
-
-		/*smallExplosion.addFrame(sf::IntRect(0, 0, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(64, 0, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(128, 0, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(192, 0, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(256, 0, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(0, 64, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(64, 64, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(128, 64, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(192, 64, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(256, 64, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(0, 128, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(64, 128, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(128, 128, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(192, 128, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(256, 128, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(0, 192, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(64, 192, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(128, 192, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(192, 192, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(256, 192, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(0, 256, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(64, 256, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(128, 256, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(192, 256, 64, 64));
-		smallExplosion.addFrame(sf::IntRect(256, 256, 64, 64));*/
 	}
 
 	if (!explosionTexture.loadFromFile("explosionSheet.png"))
@@ -159,6 +126,15 @@ AnimationManager::AnimationManager()
 
 }
 
+AnimationManager * AnimationManager::GetInstance()
+{
+	if (!instanceFlag) {
+		instance = new AnimationManager();
+		instanceFlag = true;
+	}
+	return instance;
+}
+
 void AnimationManager::Update()
 {
 	sf::Time frameTime = frameClock.restart();
@@ -166,16 +142,37 @@ void AnimationManager::Update()
 	for (int i = 0; i < animations.size(); i++)
 	{
 		animations.at(i)->update(frameTime);
+		if (animations.at(i)->isPlaying() == false)
+		{
+			checkDestroyed = true;
+		}
 	}
 
-	DeleteExpired();
+	if (checkDestroyed == true)
+	{
+		DeleteExpired();
+	}
 }
 
 void AnimationManager::Draw(sf::RenderWindow & window)
 {
 	for (int i = 0; i < animations.size(); i++)
 	{
-		window.draw(*animations.at(i));
+		if (animations.at(i)->getName() != "smokeTrail")
+		{
+			window.draw(*animations.at(i));
+		}
+	}
+}
+
+void AnimationManager::Draw2(sf::RenderWindow & window)
+{
+	for (int i = 0; i < animations.size(); i++)
+	{
+		if (animations.at(i)->getName() == "smokeTrail")
+		{
+			window.draw(*animations.at(i));
+		}
 	}
 }
 
@@ -226,6 +223,40 @@ void AnimationManager::CreateAnimation(sf::Vector2f position, float type, float 
 		animations.at(animations.size() - 1)->setRotation(rand() % 360 + 1);
 		animations.at(animations.size() - 1)->play(smallExplosion);
 	}
+	if (type == 5)
+	{
+		animations.push_back(new AnimatedSprite(sf::seconds(0.040), true, false));
+		animations.at(animations.size() - 1)->setOrigin(50, 50);
+		animations.at(animations.size() - 1)->setPosition(position);
+		float scale1 = (rand() % 5 + 1) + 20;
+		animations.at(animations.size() - 1)->setScale(sf::Vector2f(scale1 / 10, scale1 / 10));
+		//animations.at(animations.size() - 1)->setColor(sf::Color(255, 255, 255, 200));
+		animations.at(animations.size() - 1)->setRotation(rand() % 360 + 1);
+		animations.at(animations.size() - 1)->play(largeExplosion);
+	}
+	if (type == 6)
+	{
+		animations.push_back(new AnimatedSprite(sf::seconds(0.040), true, false));
+		animations.at(animations.size() - 1)->setOrigin(6, -33);
+		animations.at(animations.size() - 1)->setPosition(position);
+		float scale1 = (rand() % 5 + 1) + 20;
+		animations.at(animations.size() - 1)->setScale(sf::Vector2f(0.5, 0.5));
+		//animations.at(animations.size() - 1)->setColor(sf::Color(255, 255, 255, 100));
+		animations.at(animations.size() - 1)->setRotation(angle);
+		animations.at(animations.size() - 1)->play(smokeEffect);
+		animations.at(animations.size() - 1)->setName("smokeTrail");
+	}
+	if (type == 7)
+	{
+		animations.push_back(new AnimatedSprite(sf::seconds(0.020), true, false));
+		animations.at(animations.size() - 1)->setOrigin(50, 50);
+		animations.at(animations.size() - 1)->setPosition(position);
+		float scale1 = (rand() % 4 + 1) + 5;
+		animations.at(animations.size() - 1)->setScale(sf::Vector2f(scale1 / 10, scale1 / 10));
+		//animations.at(animations.size() - 1)->setColor(sf::Color(255, 255, 255, 200));
+		animations.at(animations.size() - 1)->setRotation(rand() % 360 + 1);
+		animations.at(animations.size() - 1)->play(smallExplosion);
+	}
 
 	DeleteExpired();
 }
@@ -244,15 +275,19 @@ void AnimationManager::DeleteExpired()
 		{
 			if (animations.size() == 1)
 			{
+				delete animations.at(0);
 				animations.clear();
 				break;
 			}
 			else
 			{
+				delete (*iter);
 				iter = animations.erase(iter);
 				iter = animations.begin();
 				endIter = animations.end();
 			}
 		}
 	}
+
+	checkDestroyed = false;
 }
